Rafactor prefix object with its own blueprints, forms, views, template
