--- webapp/forms.py
@@ -1,8 +1,0 @@
-from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField
-from wtforms.validators import DataRequired
-
-
-class PrefixForm(FlaskForm):
-    prefix = StringField('Префикс', validators=[DataRequired()])
-    submit = SubmitField('Save')
