--- webapp/templates/prefix.html
@@ -1,23 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-
-<form method="post">
-	{{ form.hidden_tag() }}
-	<p>
-		{{ form.prefix.label }}<br>
-	</p>
-	{% if form.prefix.errors %}
-	<ul class="errors">
-		{% for error in form.prefix.errors %}
-		<li>{{ error }}</li>
-		{% endfor %}
-	</ul>
-	{% endif %}
-	{{ form.prefix }}
-	<p>
-		{{ form.submit }}
-	</p>
-</form>
-
-{% endblock %}
